simple-agent/pyproject.toml
[project]
name = "simple-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-mcp-adapters>=0.1.7",
    "langchain-ollama>=0.3.3",
    "langgraph>=0.4.8",
    "python-dotenv>=1.1.0",
]
